Add hebcal_api.py
This file utilizes hebcal's API to convert back and forth between
gregorian and hebrew dates. Committing because it's working.
The next step is to start running it through the csv file.

API rate limit is 90 requests per 10 seconds.
Need to add a way to sleep the runtime if it gets that high.
It's a free API, so it's wise to be considerate.