MVP - calendar conversion works
Hebcal Api file was combined into csvconverter.py
dates.csv was removed as its no longer needed.
The process is now a csv file of hebrew dates:
hebrew_dates.csv is successfully converted into an ics file.

TODO:
the hebrew dates need to be put into the calendar's description.
Calculating the next day needs to be done with datetime so we don't
get a 32nd day.

Still a work in progress, but it works.