Add clear row, current row updates dynamically
Added a clear row button which empties out all the fields
filled out in the current table's row.

Additionally, when clicking on a different row,
the self.row updates so that you can fill out fields in different
parts of the table.
It works, but needs to be reviewed for some strange behaviors.
Need to work on a row deletion button and perhaps a box showing
showing the currently selected row.