Pyqt4 to Pyqt5
Utilized the script here: https://github.com/rferrazz/pyqt4topyqt5
to convert the Gui APP from Pyqt4 to Pyqt5.

Since the app was pretty simple at this stage, the only thing that
seemed to change was QTTableWidgetItem being a part of QtWidgets
instead of QtGui and QMainWindow is a part of QtWidgets instead of
QtGui as well.