Continue validation framework
Working on validation framework. Currently only checking
for whitepsace and none type. Need to eventually replace this
entirely with a way to validate each cell for its type.
Alternatively, each cell as its changed should be edited
to make sure the right type is going in with a warning if the
user changes it. Would this work for importing a csv though?
Maybe both could be implemented? One validation per each cell
change, and one final validation before allowing the export/conversion
of data.