Major Redesign - default secular input, no verification
This program was majorly redesigned to closer match the web based
implementation of this which you can see here: https://www.hebcal.com/yahrzeit/

Initially, the idea was to give users the option to submit either secular or
hebrew dates. However, submitting secular dates will be the primary way this will
be used and so submitting hebrew dates as a default option isn't necessary.

There still is a feature to convert a specific hebrew date to a secular date
and to then submit it to the table (Work in progress).

Additionally, by removing the ability to manually edit the table, there is no
need for verification of data. Verification only needs to be done on csv imports.

Need to prevent new rows from being created without all the fields filled in though.