Add convert hebrew date to secular
Added function to check for internet connection.
Convert to secular now actually makes the api call,
converts it to secular, fills out the text box,
fills out the table, puts the secular calendar back to the selection,
toggles back to the secular button.

Changed the name of the class from MyAPP to hebcal_converter